initial colored

+++ src/ppu/lcd.h
@@ -8,5 +8,5 @@ public:
     constexpr static uint8_t height = 144, width = 160, white = 0;
     virtual ~LCD() = default;
     virtual void refresh() = 0;
-    virtual void setPixel(uint8_t i, uint8_t j, uint8_t color) = 0;
+    virtual void setPixel(uint8_t i, uint8_t j, uint16_t color) = 0;
 };
