+++ src/sdl_lcd.h
@@ -7,9 +7,9 @@ class SDLLCD : public LCD {
 public:
     SDLLCD(SDL_Window* window);
     void refresh() override;
-    void setPixel(uint8_t i, uint8_t j, uint8_t color) override;
+    void setPixel(uint8_t i, uint8_t j, uint16_t color) override;
 private:
-    Uint32 buffer[height][width] = { 0 };
+    uint16_t buffer[height][width] = { 0 };
     SDL_Renderer* renderer;
     SDL_Texture* texture;
 };
